refactor : EmergencyPage 리팩터링 + 맵관련 유틸함수 빼내기
